Restore wg_exit mtu to 1340
Incoming large packets where not being correctly fragmented, as shown by
speedtests of 0mbps during some download tests. From what I understand
when larger packets where sent across as the first the tcp session would
simply cut off and fail.

I'm unsure as to why this was intermittent, maybe the sessions only
negotiated to the larger mtu sometimes? But it was clearly mtu related
as using iperf with a set mtu avoided it entierly.

I couldn't reproduce this issue on other exits, so something config
specific runs into this issue, as to the original issue commit "Set mtu
of wg_exit on exit side to 1500" was trying to resolve I can't reproduce
it anyhere.