Use revert strings instead of custom errors in AltheaDb
As it stands, using custom errors results in unhelpful revert strings
with no information. This makes it very hard to debug what is actually
going wrong with the transaction. By using a short revert string like
this we do not hit the contract size limit and can understand what's
going wrong.