Migrate gnosis bridge from dai to USDS
Gnosis chain (formerly xdai) has updated their bridge to use the usds
token (the successor to the dai token, which itself was the successor to
dai v1). This patch updates the token that we provide for the bridge to
be usds and adds new handling to allow auto swapping of dai (v2) if it
is deposited by the user.