Update to clarity 1.5.5, num256 0.6.0, web30 1.12.4
This patch updates the beta20 branch to the latest clarity and web30
versions which adds support for eip1559 transactions and the large
number of quality of life improvements that have been incorporated into
these libraries since beta20 was created.

This is being done mostly to handle gnosis payment failures when
updating Nethermind to the latest version (v1.35)

This also adds handling for the gnosis chain bridge USDS transition,
which is a nice bonus and hopefully the last time we'll have to support
a change like that before the move to Althea L1 is complete.

- Replace Web3::send_transaction with prepare_transaction/prepare_legacy_transaction + send_prepared_transaction
- Replace PrivateKey::from_slice with PrivateKey::from_bytes
- Remove net_version tracking (removed from web30 API)
- Update TransactionResponse access to match on enum variants (Eip1559/Legacy/Eip2930)